Ensure that images width > 1 pixel density are never upscaled
When creating a screenshot on a Retina MacBook, the resulting image files has a 2x pixel density, i.e. an element width a 100px width displayed on screen will have a width of 200px in the screenshot.

This change allows us to optionally specify a pixel density to ensure images are never upscaled (even after taking into account their pixel density).

For example, an image width a 200px width and a pixel density of 2 will have a max-width of 100px.